Fixed Broken Test
TIL that you can't expect overloads in Elixir to work like they do in
Python, especially for structs.

I had originally written a Date comparison as `end > start`, but that
doesn't apply to the `year` attribute -- it for whatever reason only
applies to month/day. The correct way to do it is `Date.compare`.

fixes #103