Fixed Field Guesser
We were getting silent errors, and I couldn't figure out for the life of
me where they were coming from.

I backed out the latest change to the module and fixed weaknesses
exposed in the tests that were added. The tests pass and local testing
of the app shows fields being added for files that are processed.

Fixes #222