Use customized preconfigured Phoenix+Webpack setup
Manually configuring Webpack was proving extremely tedious, so this
leverages the example structure in

    https://github.com/straw-hat-team/phoenix_assets_webpack

as a guideline for building our own Phoenix+Webpack pipeline. The
license from that repository is retained in /assets for later
concatenation of attribution and licenses.

Also this is predicated on a reimplementation of our custom JS, which is
currently being inserted (or even constructed) via <script> tags in
`.eex` templates. That works, and probably even has a marginal
performance benefit, but it fails to take advantage of Webpack's
features (it actually requires explicitly disabling some) and is
fairly unreadable.