Removed UniqueConstraint and DataSetDiff
I came, I saw, I tried really, really, like super extra really hard to
make this work, I failed.

I tried imposing a restriction that more or less said, _Okay data sets,
I don't like you and you don't like me, but this is how we're going to
work together for good of clients_ and part of that agreement was that
rows needed to be unique. Well, that's not how the world works. And I
knew that then, but I had hoped that providers would try to keep their
streams clean.

I did a ton of testing and wound up going 0/15 on reviewing some of our
most important and popular data sets and checking if I could reliably
enforce uniqueness on the rows.

The answer: _we are NOT responsible for dirty data_. That's part of the
current app and we're going to carry it forward. So with that in mind, I
completely gutted this project of `UniqueConstraint` and `DataSetDiff`.

All of these changes come a direct result of that, and there are more to
come in future PRs.