Streaming ETL
What was once the fevered dream of a mad man is now reality. Ladies and
gentlemen, I present you streaming downloads.

The ETL process is totally gutted: it's now a series of GenStages that
work to queue demand of ETL jobs, it handles downloading data via
streams to save on resources, and has its own process tree that doesn't
rely on poolboy (no more drops).

This also fixes a number of smaller bugs: setting timestamps on the
meta being the most annoying. I've also fixed the query that surfaces
metas ready for ingest.

Other fixes:

- metas can no longer have json source
- meta actions have been updated where necessary to accommodate proper
  tsranges in place of the hack we had in place
- controllers have been updated to use cleaner apis
- tests have been rewritten -- they're maybe a smidge faster, but that's
  nothing compared to being actually reliable

Minor Version Bumped to 0.10.0.