Simplified Ingest Pipeline
@HeyZoos was right to pick on the unnecessary overhead and complexity of
the multiple stages plus a GenServer to handle downloads. The simpler,
and better, approach is to have a single consumer hooked up to the queue
that does all the work. These can be spun up trivially (as you'll see in
the config change, you set the number and voila).

So, here's how it all works:

0. On Application Start

When the whole application fires up, the IngestQueue and IngestWorkers
start. The queue buffers demand from the workers, and it just sits there
waiting for Metas to chew on.

1. Quantum Scheduler

I have a function in PlenarioEtl that picks up Metas ready for ingest
and add them to the queue.

2. Meta Pushed to IngestQueue

When the meta is pushed to IngestQueue, it updates the `next_import`
timestamp of the Meta (adding back in the future queue for the scheduler
to pick up). It then adds the meta to its internal demand buffer queue.
When a worker is available, the meta is passed along.

3. Meta Handed off to IngestWorker

The IngestWorker now does all of the work. Rather than separating it all
out into the stages, it handles:

- touching files
- creating download file handler
- async downloading file
- passing download to ETL functions (C/TSV and Shapefile)
- updating Meta timestamps
- cleaning up the files

This is much easier to reason about now. And now that it's a single
stage we can spin up a bunch of these to work in parallel. While I
didn't do massive CSV downloads, I did test this locally as a scraper
and I was able to queue up 10 websites on 3 workers without failure.