Added Aggregating Fields for Explorer
- redid `MetaActions.get_agg_data` to always produce some usable data
- added tests for meta action

I'm running into a bunch of problems with testing still, and they seem
to center around sentry trying to pull a field from the connection
information. I've tried disabling sentry in all envs except prod, but
that doesn't help. My last resort (after spending literally 4 hours
trying to fix it) was to just comment out the test. I've confirmed that
this works in the browser under a bunch of circumstances, but apparently
the connection information in tests screws up sentry...