Add appropriately sized jumbotron image assets
In future we should automate this generation as part of the build
process, rather than keeping them in version control, but for
purposes of short-term deployment this is easier.

You can manually generate these sizes using the
`generate_breakpoint_images.sh` script in the /assets/images
directory. (It requires ImageMagick be in your PATH.)