Added New Schema to Store Metadata for Charts
The way we're currently aggregating data and rendering charts is awful.
It needed to be completely rebuilt. The new schemas, controllers, etc
allow users to create default chart renderings for their data sets that
can then be used for display in search results.

Stuff Completed:

- new `Chart` schema for top level metadata
- new `ChartDataset` schema for aggregate metadata
- controllers, views, templates to CRUD the schemas
- special endpoint for rendering the graph serverside

Stuff To Do:

- use charts in explorer
- provide api access to aggregated data (json instead of rendered html)
- add in functionality for use with leaflet
- add text fields to views

Regarding _add text fields to views_, for some charts we could really
use being able to aggregate on text fields. Leaving them off the views
made sense at first, as we only wanted to use them to filter non-text
data. But now we've come to a point where it makes sense to have it in
there.

It will be beneficial to charting, but it will also allow us to start
adding trigram indexes so we can offer basic text search capabilities as
well -- both in the explorer and via the API.

Updates #264