Added Text Fields to Data Set Views
I added text fields to the data set views to enable aggregation around
text values in charts. I also added a new index for each text field so
that we can perform full text search later. For now, we're doing nothing
with it, but later we can start using it as the indexes will be there
and populated moving forward.

This will require another migration to enable the Postgres Trigram
extension. We will also have to do manual migrations to rebuild the
views:

```elixir
ids =
  MetaActions.list(ready_only: true)
  |> Enum.map(& &1.id)

ids
|> Enum.map(&MetaActions.get/1)
|> Enum.each(&DataSetActions.down!/1)

ids
|> Enum.map(&MetaActions.get/1)
|> Enum.each(&DataSetActions.up!/1)

ids
|> Enum.map(&MetaActions.get/1)
|> Enum.each(&PlenarioEtl.import_data_set_on_demand/1)
```

Updates #264