Added Action Functions for Chart
I needed to add in functionality to limit queries to bboxes and time
ranges. It also made sense to apply aggregate granularity to the data
sets that are being aggregated by date trunc.

The unfortunate thing about Ecto is that it can be really, really dumb.
I mean part of its strength is that it's crazy versatile and you can
throw a lot at it, but the fragments don't like dynamic values, so I had
to build a billion different functions to handle granularity.

Other than that, this is pretty straight forward -- just a lot of code
to cover the cases where dynamic fragments would have been helpful.

Updates #264