Waiting for ETLs to Finish
Since the ETL Worker is a GenStage and does its work asynchronously, we
have to _wait_ on it in the tests. This unfortunately means a system
sleep. While 2 seconds seemed like enough time, apparently the servers
Travis is running are from the 90s and this times out often.

tl;dr: bumping sleep because travis is slow