Explicitly Name Indexes
We ran into an issue where data sets with long names and with
long, similarly named fields would result in naming collisions when
creating indexes. When a name exceeds a certain length, postgres
truncates it. When this happens, we wind up with naming collisions and
bring up the table/view/indexes for the data set fails.

Fixes #362