New Map Defaults, Options
The reason the maps weren't showing up on the meta detail pages was
because the defaults for the maps was never set to match the style
updates on the front end. To fix this, I swapped the value from `500` to
`50vw`.

I also added a new option to the view to differentiate between point
data coming from AoT and regular data sets. AoT point data is richer --
it has node information and links attached to it. Meta point data is
just the coordinates.

With this new option, I added a random sampling of points from the data
set to the map in the detail view. It looks fuller and more pleasant.

In order to get efficient _random_ samples of data from Postgres, I
needed to enable the `tsm_system_rows` extension that provides this kind
of functionality. There's a migration to handle this.

Updates #331