Added Maps to Search Results When No Charts
If a data set does not have charts defined for it, rather than having a
big blank space we now just throw the map of its points in.

To get this to work, I had to update the template's javascript to use
`var`s all over the place, since redeclaring stuff using `const` and
`let` is a no-no. I'm sure Ben will hammer me on this.

Fixes #331