Fixed JS Variable Scoping
So, yeah, once again I'll fess up and say I _hate_ javascript. Like
really, really hate it and want nothing to do with it. But I get forced
to use it. So when I do, I commit mortal sins in the face of the hipster
language holy, and in this case it was using the right variable
whatever-you-call-its (`var` vs `let` vs `const`). I pray that I have
atoned for my grave errors.