Do Not Show "Request Changes" to Non-Owner Users in Data Set Show
The _Request Changes_ button was being shown to all users regardless of
ownership when the data set was shown. If a non-owning user clicked the
button, it would 403.

While this is more or less the correct response, users shouldn't even be
aware of this option if they don't own the data set -- this is doubly
true for anonymous users.

This adds a simple condition to the template that ensures the user
attached to the request owns the data set -- there's a `@user_is_owner?`
variable assigned in the response anyway.

Fixes #378