Removed God Awful Explode Library
Just... wow. What an unbelievable pile of turds.

Rather than relying on a library that was clearly developed by a high
schooler, we're going to be smart about this and **just use the
functions and modules Plug already gives us**.

Plug gives us the ability to:

- set the resonse content type we want
- set the status code and message
- get an appropriate status code by either its integer value or atom
- get an appropriate message for that code
- halt further processing

We have no reason to ever consider what the client accepts via our API
because its documented to be strictly `application/json`. Clients should
only expect that content type. And overriding the `accept` content type
the client sends is a massive violation of the client/server contract.

Beyond that, as the title of the issue this solves indicates, **we do
not serve up JSON+API content**.

This whole thing went way sideways -- mainly because of good intentions
to be helpful and give clients verbose error messages. This fixes that.
We _can_ be helpful and not rely on shity anti-patterns at the same
time.

Fixes #348