Fixing Access Issue with Shim Layer
The `TsRange` struct doesn't implement the Access protocol, so by having
`meta.timerange["lower"]` and `upper` in the shim it caused a major
problem.

This wasn't surfaced during the tests because, honestly, the tests
sucked and weren't testing a _real_ data set. I shouldn't have passed
it, but oh well.

I fixed the error in the view and updated the tests to operate against
the beach lab data set.

Fixes #393