Fixed Erlang/NaiveDateTime Casting Error
The functions in TsRange are relatively robust and should be able to
handle casts from erl dates to NaiveDateTimes. Unfortunately, there were
some edge cases in the database returns that didn't include
microseconds, and in a few other situtations we had abysmally formatted
dates in data files from sources.

This covers both edges. I added two new pattern matches to the casting and
wrapped the parsing in case statements defaulting to nil.

I also suppressed a compiler warning (it was irritating me).

Fixes #395