Fix BBox Parsing
We weren't really parsing bounding boxes right -- there were some
tangles in the logic that caused the API to return 500s when it should
have recognized the problem and 400ed out.

The parsing chain was more or less correct, but it didn't surface
individual parsing errors and halt. The changes made will catch them as
they occur and then halt processing and return a 400 response.

Specific to the list endpoint, we're ensuring that the bbox
_intersects_; on the detail endpoint we only returns records whose point
is _within_.

Fixes #401