Updated API Controller Tests
While working through the fix for #401, I realized that our tests suck.
I had made changes to these before but they really just needed to be
gutted.

By doing this, I exposed a number of other bugs:

- #403
- #404
- #405
- #406
- #407
- #408

I also cross referenced this with the API docs and realized that those
aren't doing an adequate job, hence #409.

We also need to break out some of the more generic tests into their own
suite: #411 and #412.

I suspect I'll find more of the same in the AoT API and the legacy shim,
but for now I'm focusing on just the list and detail endpoints.