Detail Endpoint Now 404s for all Unknown Identifiers
So here's another fun one: what happens when you pass a legit Meta ID
instead of the slug for a ready data set to the detail endpoint?

If you said it returns the correct data set and its records as dictated
by other params, **you'd be wrong**. While this isn't the desired
behavior, you could reasonably expect that an ID would work. And if you
were wrong you'd like to know about it.

If you said it 404s because and ID is not a slug, **you'd be wrong**
also. Afterall this is what the router dictates, right?

If you said it works it way all the way down into the ModelRegistry,
finds the meta, creates and registers a dynamic model, and then explodes
because it registers both the _integer_ ID and the slug name and then
calls `Map.fetch!` with the _string_ ID, thus causing an uncaught error
and forcing a 500 response, well **unfortunately you were right**.

To fix this, I moved more code around and split out calls to private
functions that would match against errors and dump the call to
`halt_with`.

Fixes #405