Bump Minor Version: API Purge
So, I kinda sorta totally gutted and rewrote the API. I have a good
reason for doing this, I swear.

We had a handful of seemingly unrelated bugs in the API, and being the
good Orkin Man that I am I started fixing them. At first I was doing
them one at a time, but then that wasn't really fixing anything -- it
was more like handling edge cases. So then I started looking for
patterns in the bugs, and how far down the stack they could go.

The way I started this was with the tests. They sucked. I'm sorry for
being so blunt, but they really weren't that great. I probably should
have scrutinized them earlier when the PRs came up, but I think that was
back in late April/early May when I was getting married and buying a
house. Not an excuse, just some context. Or at least that's when the
bulk of the API was being developed.

Anywho, I rewrote the tests and, lo and behold, it turns out that we
were not delivering on what we promised. We also had a ton of bugs, and
some really sloppy hacks to force things together.

At the same time, I asked the CDO folks to start hammering the Shim API
and they found tons of broken stuff: the datasets endpoint didn't return
the same shape as the existing app, the fields endpoint was missing,
they didn't even try the detail endpoint but my inspection found it to
also be mishapened.

Given all of that, it seemed like giving this a fresh start was the best
approach. The controllers are a lot slimmer, the views delegate where
appropriate, the plugs are now purely functions, the utils are cleanly
laid out, and best of all the errors are all being caught and handled
properly.

I've also done away with all references to 422 status codes. I'm sorry,
that's just crazy talk. 400 is the right way to do it -- it's a bad
request.

Fixes #401
Fixes #403
Fixes #404
Fixes #405
Fixes #406
Fixes #407
Fixes #408
Fixes #415
Fixes #416
Fixes #417
Fixes #418