Cleaned and Documented API View Utils
I added docs and specs to the public functions and updated the private
`is_clean` function to be built at compile time using lists of keys and
values to define dirty types. The beauty of this is that we can just
keep updating those lists as needed without having to touch the
functions.