Adios V1 Shim
Due to the unbelievable number of bugs that were enshrined as features
in the V1 API that I absolutely refuse to reimplement in this platform,
the CoCDO devs and I decided the best course of action is to just set up
a redirect in Nginx to push V1 URIs to the legacy app.

As much as I wanted to just shut that sucker down, we can't -- for now.

Closes #424