Fix error linting Vue files; undo dependency churn
Though the docs seem to indicate linting <style> tags in HTML-like files
like Vue components works out of the box in the version of StyleLint
we're using, they don't (it tries to lint the whole Vue file, including
template and JavaScript, which goes as well as you might expect). As
such, I've added the appropriate StyleLint file processor to do it the
"old" way until the problem gets resolved.