Fixed Geom Parsing Stupidity
Once and for all, can we please get the correct order of building geoms?
I know I'm heavily to blame for this, but it needs to be put out there:

**Geoms are parsed as {x, y}**

Where x is the longitude and y is the latitude. Mnemonically everyone
thinks of location as _lat/long_ that that's not how it works in geom
libraries.

I've also done some further reading on creating bounding boxes. I knew
my original implementation was faulty and there must be a better way.
There is, and it's name is _convex hull_. That is the minimum _polygon_
that surrounds a series on geometries.

To recap:

- fixed geom parsing and creating
- fixed meta bbox computing
- fixed aot meta bbox computing

Updates #427