Added GeoJSON Formatting Option to API
Default data comes back as an array of regular JSON objects. Clients now
have the option to request result data to be formatted as an array of
GeoJSON objects.

The way this works is that we pluck whatever Geom Object we find in the
record out, then invert the record inside that geometry under its
`properties`. For example, if we have a record

```json
{
  name: "vince",
  age: "old",
  location: %Geo.Point{
    srid: 4326,
    coordinates: {1, 2}
  }
}
```

we can format that into GeoJSON as

```json
{
  type: "Feature",
  geometry: {
    type: "Point",
    coordinates: [1, 2],
    srid: 4326
  },
  properties: {
    name: "vince",
    age: "old"
  }
}
```

The obvious use case for this is to allow clients to directly plot
results on maps without having to do any of their own parsing/inverting.