Updated Counts Logic
The `meta.counts` shoudld be relevant to the response -- e.g. that when
a head request is made, we should say there is one total record, because
that's what it is. The way I wrote it originally in the branch was using
the total number of records in the data set. That's a conflation of
metadata.