Updated Data Set Export
Since we don't really have all that great of an export strategy anymore
(we've made a lot of changes to our IO workers without giving any love
to the export process), it just makes sense at the moment to open a
blank tab and send the user to the source document.

I've updated the name of the link to _Download a Copy_ rather than
exporting to make it not seem like the users are going to just get a
view of the data set that matches their query -- the ideal export
functionality.

This still needs to fixed soon, but as we are rapidly approaching
release I don't want to screw around with big stuff like this.

Updates #371