Drop AoT Tables
I added a migration that will drop the AoT tables _if they exist_. The
reason for doing that is because I also commented out all the migration
steps that created and altered AoT tables.

So, if someone is running the migrations for the first time the AoT
tables will never be created. If you're running this with this being the
only change it will drop the tables.

Updates #458