Toward a sane internal API and Socrata native ingests
The internal API was getting really nasty -- we had a bunch of one off
functions that clashed in arity (positional arguments, matches, guards,
options ...).

The web application was also a disaster -- originally I thought it would
make it easier to keep the web, admin and API separate in subapps,
but that ended up making things that much more difficult.

Then that leaves the elephant in the room: Socrata. We've always relied
on them and all of their awful decisions. The changes here in remove
some of the terrible things about Socrata integration and makes
ingesting their data sets a little cleaner.

Breaking Changes:

- total revision of the migrations
- entirely removed the `UserAdminMessage` schema
- entirely removes all the outstanding ETL job stuff
- entirely removes charts -- that was a really stupid idea
- entirely removes exports -- again just a stupid idea
- totally new ingest pipeline
- slimmed down the API (still needs some work0

Closes #235
Closes #340
Closes #360
Closes #361