Better deployment strategy
For a _really_ long time now I've wanted a better build and deployment
strategy. This fixes a lot of the pain involved with that.

At the top level of all of this is a new Makefile. To build and deploy a
new release all you need to run is `make deploy`. See the readme file
for more details on deployment options.

Under the hood I've done a few things:

1. retooled the Dockerfile to build a base image
1. updated the build script to use the new base Docker image to build a
   release archive
1. updated the deploy script to be more efficient and utilize the new
   `seed` binary in the release
1. moved those script to a new _bin_ directory

I've also corrected the .tool-versions file, updated .dockerignore and
.gitignore and added back Sentry.